Cleanup

+++ codec/upickle/src/main/scala/jsonrpc/codec/common/upickle/UpickleCustom.scala
@@ -32,8 +32,8 @@ trait UpickleCustom extends AttributeTagged:
       override def expectedMsg = "expected number"
       override def visitString(s: CharSequence, index: Int) = s.toString.toDouble
       override def visitInt32(d: Int, index: Int) = d
-      override def visitInt64(d: Long, index: Int) = d
-      override def visitUInt64(d: Long, index: Int) = d
+      override def visitInt64(d: Long, index: Int) = d.toDouble
+      override def visitUInt64(d: Long, index: Int) = d.toDouble
       override def visitFloat64(d: Double, index: Int) = d
 
       override def visitFloat64StringParts(s: CharSequence, decIndex: Int, expIndex: Int, index: Int) =
